bootutil: Refactor encrypted.c to have the same logic as image_validate
Currently encrypted.c contains all the implementation of the
functions inside enc_key.h for each configuration (RSA, EC256
and X25519).

The purpose of this pr is to adopt a logic similar to the
file organization of image_validate to make it easier to
add other confifurations.
